chore: command bar refactor of search result items for consistent styling and icons
