+++ frontend/src/component/commandBar/RecentlyVisited/CommandResultGroup.tsx
@@ -10,30 +10,43 @@ import {
 import { Link } from 'react-router-dom';
 import type { Theme } from '@mui/material/styles/createTheme';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
-import { StyledProjectIcon } from 'component/layout/MainLayout/NavigationSidebar/IconRenderer';
+import {
+    IconRenderer,
+    StyledProjectIcon,
+} from 'component/layout/MainLayout/NavigationSidebar/IconRenderer';
 import { TooltipResolver } from 'component/common/TooltipResolver/TooltipResolver';
 import { usePlausibleTracker } from 'hooks/usePlausibleTracker';
+import useProjectOverview from 'hooks/api/getters/useProjectOverview/useProjectOverview';
 
-const listItemButtonStyle = (theme: Theme) => ({
-    borderRadius: theme.spacing(0.5),
+export const listItemButtonStyle = (theme: Theme) => ({
+    border: `1px solid transparent`,
     borderLeft: `${theme.spacing(0.5)} solid transparent`,
-    '&.Mui-selected': {
+    '&:hover, &:focus': {
+        border: `1px solid ${theme.palette.primary.main}`,
         borderLeft: `${theme.spacing(0.5)} solid ${theme.palette.primary.main}`,
     },
 });
+const StyledContainer = styled('div')(({ theme }) => ({
+    marginBottom: theme.spacing(3),
+}));
 
-const StyledTypography = styled(Typography)(({ theme }) => ({
+export const StyledTypography = styled(Typography)(({ theme }) => ({
     fontSize: theme.fontSizes.bodySize,
-    padding: theme.spacing(0, 3),
+    padding: theme.spacing(0, 2.5),
 }));
 
-const StyledListItemIcon = styled(ListItemIcon)(({ theme }) => ({
-    minWidth: theme.spacing(4),
-    margin: theme.spacing(0.25, 0),
+export const StyledListItemIcon = styled(ListItemIcon)(({ theme }) => ({
+    minWidth: theme.spacing(0.5),
+    margin: theme.spacing(0, 1, 0, 0),
 }));
 
-const StyledListItemText = styled(ListItemText)(({ theme }) => ({
+export const StyledListItemText = styled(ListItemText)(({ theme }) => ({
     margin: 0,
+    fontSize: theme.fontSizes.bodySize,
+}));
+
+export const StyledButtonTypography = styled(Typography)(({ theme }) => ({
+    fontSize: theme.fontSizes.bodySize,
 }));
 
 export interface CommandResultGroupItem {
@@ -42,23 +55,149 @@ export interface CommandResultGroupItem {
     description?: string | null;
 }
 
+export const RecentlyVisitedPathButton = ({
+    path,
+    key,
+    name,
+}: { path: string; key: string; name: string }) => {
+    const { trackEvent } = usePlausibleTracker();
+
+    const onClick = () => {
+        trackEvent('command-bar', {
+            props: {
+                eventType: `click`,
+                source: 'recently-visited',
+                eventTarget: 'Pages',
+                pageType: name,
+            },
+        });
+    };
+
+    return (
+        <ListItemButton
+            key={key}
+            dense={true}
+            component={Link}
+            to={path}
+            sx={listItemButtonStyle}
+            onClick={onClick}
+        >
+            <StyledListItemIcon>
+                <ConditionallyRender
+                    condition={path === '/projects'}
+                    show={<StyledProjectIcon />}
+                    elseShow={<IconRenderer path={path} />}
+                />
+            </StyledListItemIcon>
+            <StyledListItemText>
+                <StyledButtonTypography color='textPrimary'>
+                    {name}
+                </StyledButtonTypography>
+            </StyledListItemText>
+        </ListItemButton>
+    );
+};
+
+export const RecentlyVisitedProjectButton = ({
+    projectId,
+    key,
+}: { projectId: string; key: string }) => {
+    const { trackEvent } = usePlausibleTracker();
+    const { project, loading } = useProjectOverview(projectId);
+    const projectDeleted = !project.name && !loading;
+
+    const onClick = () => {
+        trackEvent('command-bar', {
+            props: {
+                eventType: `click`,
+                source: 'recently-visited',
+                eventTarget: 'Projects',
+            },
+        });
+    };
+
+    if (projectDeleted) return null;
+    return (
+        <ListItemButton
+            key={key}
+            dense={true}
+            component={Link}
+            to={`/projects/${projectId}`}
+            sx={listItemButtonStyle}
+            onClick={onClick}
+        >
+            <StyledListItemIcon>
+                <StyledProjectIcon />
+            </StyledListItemIcon>
+            <StyledListItemText>
+                <StyledButtonTypography color='textPrimary'>
+                    {project.name}
+                </StyledButtonTypography>
+            </StyledListItemText>
+        </ListItemButton>
+    );
+};
+
+export const RecentlyVisitedFeatureButton = ({
+    key,
+    projectId,
+    featureId,
+}: {
+    key: string;
+    projectId: string;
+    featureId: string;
+}) => {
+    const onClick = () => {
+        const { trackEvent } = usePlausibleTracker();
+
+        trackEvent('command-bar', {
+            props: {
+                eventType: `click`,
+                source: 'recently-visited',
+                eventTarget: 'Flags',
+            },
+        });
+    };
+    return (
+        <ListItemButton
+            key={key}
+            dense={true}
+            component={Link}
+            to={`/projects/${projectId}/features/${featureId}`}
+            sx={listItemButtonStyle}
+            onClick={onClick}
+        >
+            <StyledListItemIcon>
+                <Icon>{'flag'}</Icon>
+            </StyledListItemIcon>
+            <StyledListItemText>
+                <StyledButtonTypography color='textPrimary'>
+                    {featureId}
+                </StyledButtonTypography>
+            </StyledListItemText>
+        </ListItemButton>
+    );
+};
+
 interface CommandResultGroupProps {
     icon: string;
     groupName: string;
-    items: CommandResultGroupItem[];
+    items?: CommandResultGroupItem[];
+    children?: React.ReactNode;
 }
 
 export const CommandResultGroup = ({
     icon,
     groupName,
     items,
+    children,
 }: CommandResultGroupProps) => {
     const { trackEvent } = usePlausibleTracker();
-    const slicedItems = items.slice(0, 3);
-
-    if (items.length === 0) {
+    if (!children && (!items || items.length === 0)) {
         return null;
     }
+
+    const slicedItems = items?.slice(0, 3);
 
     const onClick = (item: CommandResultGroupItem) => {
         trackEvent('command-bar', {
@@ -70,13 +209,15 @@ export const CommandResultGroup = ({
             },
         });
     };
+
     return (
-        <>
+        <StyledContainer>
             <StyledTypography color='textSecondary'>
                 {groupName}
             </StyledTypography>
             <List>
-                {slicedItems.map((item, index) => (
+                {children}
+                {slicedItems?.map((item, index) => (
                     <ListItemButton
                         key={`command-result-group-${groupName}-${index}`}
                         dense={true}
@@ -100,12 +241,14 @@ export const CommandResultGroup = ({
                             placement={'bottom-end'}
                         >
                             <StyledListItemText>
-                                <Typography>{item.name}</Typography>
+                                <StyledButtonTypography color='textPrimary'>
+                                    {item.name}
+                                </StyledButtonTypography>
                             </StyledListItemText>
                         </TooltipResolver>
                     </ListItemButton>
                 ))}
             </List>
-        </>
+        </StyledContainer>
     );
 };
